separated CLI from sensors
